rpiRobot/requirements.txt
colorzero==1.1
coverage==4.5.2
gpiozero==1.5.0
nose==1.3.7
numpy==1.16.1
opencv-python==3.4.4.19
pyserial==3.4
